% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{input_nas}
\alias{input_nas}
\title{Input NAs in a data.table with the column's median, mean or zeros}
\usage{
Notice that \code{dt} will be modified dinamycally with the ':='
operator. Make sure you make a hard copy of \code{dt} if you want to keep the
original data.table
}
\arguments{
\item{dt}{Data table}

\item{cols}{Columns from \code{dt}}

\item{treatment}{Accepts "zero", "mean", "median"}
}
\value{
Dinamically transforms \code{dt} with columns in \code{vars}
treated to input NA's
}
\description{
Input NAs in a data.table with the column's median, mean or zeros
}
