% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{print_truth}
\alias{print_truth}
\title{A test function that prints "Lucho es el mejor!" by default,
but can also print someone else's name}
\usage{
By default, it prints the truth, but you can change \code{name}
to be nice to someone else
}
\arguments{
\item{name}{Name of the person who rox}
}
\value{
String saying who rox
}
\description{
A test function that prints "Lucho es el mejor!" by default,
but can also print someone else's name
}
